@@ -10,14 +10,11 @@ keywords =
   phantom
 classifiers =
   Development Status :: 5 - Production/Stable
-  Environment :: Console\
+  Environment :: Console
   Intended Audience :: Developers
   Intended Audience :: End Users/Desktop
   Operating System :: OS Independent
   Programming Language :: Python :: 3
-  Programming Language :: Python :: 3.6
-  Programming Language :: Python :: 3.7
-  Programming Language :: Python :: 3.8
   Topic :: Utilities
 license_files =
   LICENSE.txt
@@ -25,11 +22,9 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 
 [options]
-python_requires = >= 3.6
+python_requires = >= 3.7
 packages = find:
-zip_safe: false
-scripts =
-
+zip_safe = False
 install_requires =
   numpy
   scipy
@@ -45,7 +40,3 @@ lint =
 io =
   imageio
   imageio-ffmpeg
-
-
-[options.entry_points]
-console_scripts =
